requirements.txt
Django>=2.2.0
djangorestframework>=3.11.0
django-filter == 2.2.0
django-cors-headers
django-crispy-forms==1.6.0
django-betterforms == 1.2.0
djangorestframework-simplejwt == 4.4.0

